Update world.py
This is what I have so far for the Special Room Tile and the Found Item tile. Please note that anything can be changed and that I think it needs some tweaking for it to run properly. I also have a comment in here stating that I am not completely done with the Found Item tile. I am working on the code to increment the player's credit by the value of the item they find.